+++ test_tools/log_parser/binary_parser.cc
@@ -1,67 +1,10 @@
 #include <boost/program_options.hpp>
 #include <cstring>
-#include <fstream>
 #include <iostream>
-#include <vector>
-
+#include <string>
 #include "joined_log_parser.h"
-#include "example_joiner.h"
 
 namespace po = boost::program_options;
-
-// TODO: what assumptions should we make about endianness?
-// TODO make better error messages
-int read_and_deserialize_file(const std::string &file_name) {
-  std::ifstream fs(file_name.c_str(), std::ifstream::binary);
-  std::vector<char> buffer(4, 0);
-  const std::vector<char> magic = {'V', 'W', 'F', 'B'};
-
-  // read the 4 magic bytes
-  fs.read(buffer.data(), buffer.size());
-  if (buffer != magic) {
-    return err::file_read_error;
-  }
-
-  // read the version
-  fs.read(buffer.data(), buffer.size());
-  if (static_cast<int>(buffer[0]) != 1) {
-    return err::file_read_error;
-  }
-
-  // payload type, check for header
-  unsigned int payload_type;
-  fs.read((char *)(&payload_type), sizeof(payload_type));
-  if (payload_type != MSG_TYPE_HEADER) {
-    return err::file_read_error;
-  }
-
-  // read header size
-  fs.read(buffer.data(), buffer.size());
-  uint32_t payload_size = *reinterpret_cast<const uint32_t *>(buffer.data());
-  // read the payload
-  std::vector<char> payload(payload_size, 0);
-  fs.read(payload.data(), payload.size());
-  // data is in the payload now can use it to deserialize header
-  JoinedLogParser logparser("--quiet --cb_explore_adf -f amodel.model");
-  logparser.read_header(payload);
-
-  fs.read((char *)(&payload_type), sizeof(payload_type));
-  while (payload_type != MSG_TYPE_EOF) {
-    if (payload_type != MSG_TYPE_REGULAR)
-      return err::file_read_error;
-    // read payload size
-    fs.read(buffer.data(), buffer.size());
-    uint32_t payload_size = *reinterpret_cast<const uint32_t *>(buffer.data());
-    // read the payload
-    std::vector<char> payload(payload_size, 0);
-    fs.read(payload.data(), payload.size());
-    if (logparser.read_message(payload) != err::success)
-      return err::file_read_error;
-    fs.read((char *)(&payload_type), sizeof(payload_type));
-  }
-
-  return err::success;
-}
 
 int main(int argc, char *argv[]) {
   po::options_description desc("binary file parser");
